Update .travis.yml
For reasons unknown, the remove directory command is failing even though the ls command (when tests are executing) clearly shows that the pyglow directory exists. Trying a different approach this time, using the variable for the pyglow directory. Also added commands to list what is going on in the directory..... This Travis CI cache issue is costing more time than it saves.